se agrego en pagar venta mayoreo la opcion para pagar venta menudeo

+++ src/main/java/com/web/chon/negocio/NegocioBuscaVenta.java
@@ -1,11 +1,7 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package com.web.chon.negocio;
 
-import com.web.chon.dominio.Cliente;
+import java.math.BigDecimal;
 import java.util.List;
 import javax.ejb.Remote;
 
@@ -29,7 +25,7 @@ public interface NegocioBuscaVenta
      * Busca una venta por medio de un folio y el id de sucursal
      * 
      */
-    public List<Object[]> getVentaMenudeoByfolioAndIdSuc(int folio, int idSucursal);
+    public List<Object[]> getVentaMenudeoByfolioAndIdSuc(BigDecimal folio, int idSucursal);
     
 
 }
